conversion warning fix

+++ src/mtsmsme/processor/ACRepo.cpp
@@ -49,7 +49,7 @@ string TrId::toString()
   string AC::toString()
   {
     int k = 0;
-    int len = arcs.size();
+    int len = (int)arcs.size();
     char text[32] = {0,};
     std::vector<unsigned char> stream;
 
@@ -63,52 +63,52 @@ string TrId::toString()
     return result;
   }
 static unsigned long           null_buf[] = {0,0,0,0,0,0, 0,0};
-AC null_ac        = AC(null_buf,sizeof(null_buf)/sizeof(unsigned long));
+AC null_ac        = AC(null_buf,(int)(sizeof(null_buf)/sizeof(unsigned long)));
 
 static unsigned long sm_mt_relay_v1_buf[] = {0,4,0,0,1,0,25,1};
 static unsigned long sm_mt_relay_v2_buf[] = {0,4,0,0,1,0,25,2};
 static unsigned long sm_mt_relay_v3_buf[] = {0,4,0,0,1,0,25,3};
-AC sm_mt_relay_v1 = AC(sm_mt_relay_v1_buf,sizeof(sm_mt_relay_v1_buf)/sizeof(unsigned long));
-AC sm_mt_relay_v2 = AC(sm_mt_relay_v2_buf,sizeof(sm_mt_relay_v2_buf)/sizeof(unsigned long));
-AC sm_mt_relay_v3 = AC(sm_mt_relay_v3_buf,sizeof(sm_mt_relay_v3_buf)/sizeof(unsigned long));
+AC sm_mt_relay_v1 = AC(sm_mt_relay_v1_buf,(int)(sizeof(sm_mt_relay_v1_buf)/sizeof(unsigned long)));
+AC sm_mt_relay_v2 = AC(sm_mt_relay_v2_buf,(int)(sizeof(sm_mt_relay_v2_buf)/sizeof(unsigned long)));
+AC sm_mt_relay_v3 = AC(sm_mt_relay_v3_buf,(int)(sizeof(sm_mt_relay_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long net_loc_upd_v1_buf[] = {0,4,0,0,1,0, 1,1};
 static unsigned long net_loc_upd_v2_buf[] = {0,4,0,0,1,0, 1,2};
 static unsigned long net_loc_upd_v3_buf[] = {0,4,0,0,1,0, 1,3};
-AC net_loc_upd_v1 = AC(net_loc_upd_v1_buf,sizeof(net_loc_upd_v1_buf)/sizeof(unsigned long));
-AC net_loc_upd_v2 = AC(net_loc_upd_v2_buf,sizeof(net_loc_upd_v2_buf)/sizeof(unsigned long));
-AC net_loc_upd_v3 = AC(net_loc_upd_v3_buf,sizeof(net_loc_upd_v3_buf)/sizeof(unsigned long));
+AC net_loc_upd_v1 = AC(net_loc_upd_v1_buf,(int)(sizeof(net_loc_upd_v1_buf)/sizeof(unsigned long)));
+AC net_loc_upd_v2 = AC(net_loc_upd_v2_buf,(int)(sizeof(net_loc_upd_v2_buf)/sizeof(unsigned long)));
+AC net_loc_upd_v3 = AC(net_loc_upd_v3_buf,(int)(sizeof(net_loc_upd_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long net_loc_cancel_v1_buf[] = {0,4,0,0,1,0,2,1};
 static unsigned long net_loc_cancel_v2_buf[] = {0,4,0,0,1,0,2,2};
 static unsigned long net_loc_cancel_v3_buf[] = {0,4,0,0,1,0,2,3};
-AC net_loc_cancel_v1 = AC(net_loc_cancel_v1_buf,sizeof(net_loc_cancel_v1_buf)/sizeof(unsigned long));
-AC net_loc_cancel_v2 = AC(net_loc_cancel_v2_buf,sizeof(net_loc_cancel_v2_buf)/sizeof(unsigned long));
-AC net_loc_cancel_v3 = AC(net_loc_cancel_v3_buf,sizeof(net_loc_cancel_v3_buf)/sizeof(unsigned long));
+AC net_loc_cancel_v1 = AC(net_loc_cancel_v1_buf,(int)(sizeof(net_loc_cancel_v1_buf)/sizeof(unsigned long)));
+AC net_loc_cancel_v2 = AC(net_loc_cancel_v2_buf,(int)(sizeof(net_loc_cancel_v2_buf)/sizeof(unsigned long)));
+AC net_loc_cancel_v3 = AC(net_loc_cancel_v3_buf,(int)(sizeof(net_loc_cancel_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long shortMsgGatewayContext_v1_buf[] = {0,4,0,0,1,0,20,1};
 static unsigned long shortMsgGatewayContext_v2_buf[] = {0,4,0,0,1,0,20,2};
 static unsigned long shortMsgGatewayContext_v3_buf[] = {0,4,0,0,1,0,20,3};
-AC shortMsgGatewayContext_v1 = AC(shortMsgGatewayContext_v1_buf,sizeof(shortMsgGatewayContext_v1_buf)/sizeof(unsigned long));
-AC shortMsgGatewayContext_v2 = AC(shortMsgGatewayContext_v2_buf,sizeof(shortMsgGatewayContext_v2_buf)/sizeof(unsigned long));
-AC shortMsgGatewayContext_v3 = AC(shortMsgGatewayContext_v3_buf,sizeof(shortMsgGatewayContext_v3_buf)/sizeof(unsigned long));
+AC shortMsgGatewayContext_v1 = AC(shortMsgGatewayContext_v1_buf,(int)(sizeof(shortMsgGatewayContext_v1_buf)/sizeof(unsigned long)));
+AC shortMsgGatewayContext_v2 = AC(shortMsgGatewayContext_v2_buf,(int)(sizeof(shortMsgGatewayContext_v2_buf)/sizeof(unsigned long)));
+AC shortMsgGatewayContext_v3 = AC(shortMsgGatewayContext_v3_buf,(int)(sizeof(shortMsgGatewayContext_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long roamingNumberEnquiryContext_v1_buf[] = {0,4,0,0,1,0,3,1};
 static unsigned long roamingNumberEnquiryContext_v2_buf[] = {0,4,0,0,1,0,3,2};
 static unsigned long roamingNumberEnquiryContext_v3_buf[] = {0,4,0,0,1,0,3,3};
-AC roamingNumberEnquiryContext_v1 = AC(roamingNumberEnquiryContext_v1_buf,sizeof(roamingNumberEnquiryContext_v1_buf)/sizeof(unsigned long));
-AC roamingNumberEnquiryContext_v2 = AC(roamingNumberEnquiryContext_v2_buf,sizeof(roamingNumberEnquiryContext_v2_buf)/sizeof(unsigned long));
-AC roamingNumberEnquiryContext_v3 = AC(roamingNumberEnquiryContext_v3_buf,sizeof(roamingNumberEnquiryContext_v3_buf)/sizeof(unsigned long));
+AC roamingNumberEnquiryContext_v1 = AC(roamingNumberEnquiryContext_v1_buf,(int)(sizeof(roamingNumberEnquiryContext_v1_buf)/sizeof(unsigned long)));
+AC roamingNumberEnquiryContext_v2 = AC(roamingNumberEnquiryContext_v2_buf,(int)(sizeof(roamingNumberEnquiryContext_v2_buf)/sizeof(unsigned long)));
+AC roamingNumberEnquiryContext_v3 = AC(roamingNumberEnquiryContext_v3_buf,(int)(sizeof(roamingNumberEnquiryContext_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long subscriberDataMngtContext_v1_buf[] = {0,4,0,0,1,0,16,1};
 static unsigned long subscriberDataMngtContext_v2_buf[] = {0,4,0,0,1,0,16,2};
 static unsigned long subscriberDataMngtContext_v3_buf[] = {0,4,0,0,1,0,16,3};
-AC subscriberDataMngtContext_v1 = AC(subscriberDataMngtContext_v1_buf,sizeof(subscriberDataMngtContext_v1_buf)/sizeof(unsigned long));
-AC subscriberDataMngtContext_v2 = AC(subscriberDataMngtContext_v2_buf,sizeof(subscriberDataMngtContext_v2_buf)/sizeof(unsigned long));
-AC subscriberDataMngtContext_v3 = AC(subscriberDataMngtContext_v3_buf,sizeof(subscriberDataMngtContext_v3_buf)/sizeof(unsigned long));
+AC subscriberDataMngtContext_v1 = AC(subscriberDataMngtContext_v1_buf,(int)(sizeof(subscriberDataMngtContext_v1_buf)/sizeof(unsigned long)));
+AC subscriberDataMngtContext_v2 = AC(subscriberDataMngtContext_v2_buf,(int)(sizeof(subscriberDataMngtContext_v2_buf)/sizeof(unsigned long)));
+AC subscriberDataMngtContext_v3 = AC(subscriberDataMngtContext_v3_buf,(int)(sizeof(subscriberDataMngtContext_v3_buf)/sizeof(unsigned long)));
 
 static unsigned long shortMsgMoRelayContext_v2_buf[] = {0,4,0,0,1,0,21,2};
-AC shortMsgMoRelayContext_v2 = AC(shortMsgMoRelayContext_v2_buf,sizeof(shortMsgMoRelayContext_v2_buf)/sizeof(unsigned long));
+AC shortMsgMoRelayContext_v2 = AC(shortMsgMoRelayContext_v2_buf,(int)(sizeof(shortMsgMoRelayContext_v2_buf)/sizeof(unsigned long)));
 
 bool isMapV1ContextSupported(AC& appcntx)
 {
